AIService/requirements.txt
fastapi
uvicorn
kafka-python
pymongo
sentence-transformers
torch
python-dotenv